@@ -1,18 +1,25 @@
-const path = require('path')
-const merge = require('deepmerge')
-
 module.exports = {
-  configureWebpack: {
-    resolve: {
-      symlinks: false
-    }
-  },
-
   chainWebpack: config => {
+    config.resolve.alias.set('vue', '@vue/compat')
+
+    config.module
+      .rule('vue')
+      .use('vue-loader')
+      .tap(options => {
+        return {
+          ...options,
+          compilerOptions: {
+            compatConfig: {
+              MODE: 2
+            }
+          }
+        }
+      })
+
     config.module
       .rule('md')
       .test(/\.md$/)
-        .use('raw-loader')
-        .loader('raw-loader')
+      .use('raw-loader')
+      .loader('raw-loader')
   }
 }
